refactor: hexagonal directory
