+++ generators/domain/main.go
@@ -6,7 +6,8 @@ import (
 	"fmt"
 	"os"
 
-	 "github.com/rapidstellar/gohexa/generators/utils"
+	"github.com/rapidstellar/gohexa/internal/core/domain"
+	"github.com/rapidstellar/gohexa/internal/core/services"
 )
 
 func main() {
@@ -31,6 +32,9 @@ func main() {
 		fmt.Printf("Error creating directories: %v\n", err)
 		return
 	}
-
-	utils.GenerateDomainFile(*outputDir, *featureName, *projectName, *useUUID)
+	srv := services.NewGeneratorService(domain.GeneratorFlagDomain{
+		FeatureName: *featureName,
+		ProjectName: *projectName,
+	})
+	srv.GenerateDomainFile(*outputDir, *useUUID)
 }
