+++ pkgs/configs/constants.go
@@ -1,0 +1,5 @@
+package configs
+
+var (
+	TEMPLATE_URL = "https://github.com/rapidstellar/gohexa/releases/download/1.0.0/templates.zip"
+)
